# openapi/openapi-api.yaml
openapi: 3.1.0
info:
  title: Multi Language Demo API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Java Spring Boot service
  - url: http://localhost:8000
    description: Python FastAPI service
jsonSchemaDialect: https://json-schema.org/draft/2020-12/schema
paths:
  /messages/java:
    get:
      summary: List messages seen by Java
      operationId: listJavaMessages
      responses:
        '200':
          description: Messages stored by Java service
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoMessage'
    post:
      summary: Publish a message through Java service
      operationId: publishJavaMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemoMessage'
      responses:
        '202':
          description: Message accepted and forwarded to Kafka
  /messages/java/from-python:
    get:
      summary: Messages originally emitted by Python
      operationId: listMessagesFromPython
      responses:
        '200':
          description: Consumed records from Python topic
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoMessage'
  /messages/python:
    get:
      summary: List messages seen by Python
      operationId: listPythonMessages
      responses:
        '200':
          description: Messages stored by Python service
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoMessage'
    post:
      summary: Publish a message through Python service
      operationId: publishPythonMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemoMessage'
      responses:
        '202':
          description: Message accepted and forwarded to Kafka
  /messages/python/from-java:
    get:
      summary: Messages produced by Java service
      operationId: listMessagesFromJava
      responses:
        '200':
          description: Consumed records from Java topic
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoMessage'
components:
  schemas:
    DemoMessage:
      $ref: ./schemas/demo-message.schema.json